chore: moved ConfigRepository.kt to config directory
